Add Delete Log. Verify if a file exists before try to delete

+++ services/archivist/archivist/src/main/kotlin/com/zorroa/archivist/storage/AwsProjectStorageService.kt
@@ -15,6 +15,7 @@ import com.amazonaws.services.s3.model.ObjectMetadata
 import com.amazonaws.services.s3.model.PutObjectRequest
 import com.google.cloud.storage.StorageException
 import com.zorroa.archivist.domain.ArchivistException
+import com.zorroa.archivist.domain.AssetSpec
 import com.zorroa.archivist.domain.FileStorage
 import com.zorroa.archivist.domain.ProjectStorageLocator
 import com.zorroa.archivist.domain.ProjectStorageSpec
@@ -133,9 +134,12 @@ class AwsProjectStorageService constructor(
 
     override fun delete(locator: ProjectStorageLocator) {
         try {
-            val s3Object = s3Client.getObject(GetObjectRequest(properties.bucket, locator.getPath()))
-            var obj = DeleteObjectRequest(properties.bucket, s3Object.key)
-            s3Client.deleteObject(obj)
+            if (s3Client.doesObjectExist(properties.bucket, locator.getPath())) {
+                val s3Object = s3Client.getObject(GetObjectRequest(properties.bucket, locator.getPath()))
+                var obj = DeleteObjectRequest(properties.bucket, s3Object.key)
+                s3Client.deleteObject(obj)
+                logDeleteEvent(locator)
+            }
         } catch (ex: AmazonS3Exception) {
             throw ArchivistException(ex)
         }
